Split code to be tested into static library
Google gtest framework works by linking against a static library or dll.
Previously was having to rebuild the project as a static library when
wanting to run the unit tests. Now the main application links against
the static library containing the code to be tested. The test project
also links against this static library meaning the main app does not
need to be rebuilt as a lib.