Add better handling for XML object loading
The application would not give an error message when loading an object
XML file that referenced a model file that did not exist. In this
situation the object would not appear in the scene. I have now added
handling so that the XML load will fail if the model file referenced in
the XML definition does not exist.