Refactor game engine and camera code
Camera was storing mouse pos and view mode. The view mode included
selection which the camera couldn't really handle. Also required screen
size to be passed to camera which is a bit dodgy. Moved this logic to
game engine then camera is only concerned with what to do when panning,
zooming and rotating.